requirements.txt
aiogram==2.25.1
openai==1.6.1
supabase==2.0.3
python-dotenv==1.0.0
requests==2.31.0 